Remove magic template vars (closes #2)
This removes the magic template vars that were used in `js/*` files to
refer to dynamically created module files.

Instead, all dynamic data (slides, layouts, etc) is now added to a single
module, `js/config.js`, which is auto-generated inside
`(src|lib)/index.js`. (I realize the name of the file is not the most
appropriate).
Modules that need access to layouts or the slides, can simply import
`js/config`.

This means we don't have to pass every JS file through `_.template`
anymore, which speeds up the bundle process.

Note that static files are still passed through `_.template`, so you can
still refer to the `meta` values defined in the config, via `meta` now
instead of `META`.

If you need to refer to `meta` data in JS files, you could load the
`package.json` file via `fs.readFileSync`. We are still going to use the
`brfs` transform to inline files.

Overall this makes it easier to follow the generation process.