Improve accessibility
This makes the links in the menu non-tabbable, so that is easier to
tab between menu and slide.

While in the menu, `<Tab>` will bring you to the slide.
While on the slide, `<Shit-Tab>` will bring you to the menu.

The left and right arrow keys allow to navigate through the presentation
and will focus to current menu item.